2d API_18
monster AI 기본 세팅 및 IDLE 구현
눈물 데미지 몬스터에게 적용되게
눈물의 데미지가 생성자의 데미지에 따라 달라지게 적용
피격 애니메이션 적용중

+++ 20220302/Core.cpp
@@ -45,7 +45,8 @@ int CCore::init(HWND _hWnd, Vec2 _vResolution)
 
 
 	
-	RECT rt = { 0.f, 0.f, m_vResolution.x, m_vResolution.y };
+	RECT rt = { static_cast<long>(0.f), static_cast<long>(0.f)
+		, static_cast<long>(m_vResolution.x), static_cast<long>(m_vResolution.y) };
 	
 	
 	AdjustWindowRect(&rt, WS_OVERLAPPEDWINDOW, true);
